chore: change files order
