+++ projects/demo/src/pages/kit/time/examples/3-separators/mask.ts
