+++ projects/demo/src/pages/kit/time/examples/7-time-segments-min-max/mask.ts
